Comments: Add support for authentication token
If provided, a Mastodon application access token with the
`read:statuses` scope will enable fetching more than 60 replies to a
post.

NOTE: IF INCLUDED, ANYONE WILL BE ABLE TO READ THE ASSOCIATED ACCOUNT'S
PRIVATE STATUSES. It is highly recommended to use a dedicated bot/API
account to create an application with scope `read:statuses`.